+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Please, check out guidelines: https://keepachangelog.com/en/1.0.0/
 * Documentation to `xcbuddykit` models.
 * Equatable conformance to `xcbuddykit` models.
 * Circular dependency detection.
+* `Shell` util https://github.com/xcbuddy/xcbuddy/pull/15 by @pepibumur.
 
 ### Changed